third_party/WebKit/LayoutTests/fast/css/getComputedStyle/computed-style-redistribution-expected.txt

CONSOLE WARNING: line 17: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Check that getComputedStyle causes a shadow re-distribution when necessary.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(d1).color is white
PASS getComputedStyle(d2).color is green
PASS getComputedStyle(d1).color is green
PASS getComputedStyle(d2).color is white
PASS successfullyParsed is true

TEST COMPLETE

A
